move files to more relevant places
