+++ src/cow-react/modules/swap/services/ethFlow/index.ts
@@ -2,8 +2,8 @@ import { EthFlowContext } from '@cow/modules/swap/services/swapFlow/types'
 import { swapFlowAnalytics } from '@cow/modules/swap/services/swapFlow/steps/analytics'
 import { signEthFlowOrderStep } from '@cow/modules/swap/services/ethFlow/steps/signEthFlowOrderStep'
 import confirmPriceImpactWithoutFee from 'components/swap/confirmPriceImpactWithoutFee'
-import { logSwapFlow } from '@cow/modules/swap/services/swapFlow/logger'
-import { getSwapErrorMessage } from '@cow/modules/swap/services/swapFlow/steps/swapErrorHelper'
+import { logSwapFlow } from '@cow/modules/swap/services/utils/logger'
+import { getSwapErrorMessage } from '@cow/modules/swap/services/common/steps/swapErrorHelper'
 import { PriceImpact } from 'hooks/usePriceImpact'
 import { addPendingEthFLowOrderStep } from '@cow/modules/swap/services/ethFlow/steps/addPendingEthFLowOrderStep'
