+++ src/cow-react/modules/swap/services/swapFlow/steps/presignOrderStep.ts
@@ -2,7 +2,7 @@ import { calculateGasMargin } from 'utils/calculateGasMargin'
 import { GPv2Settlement } from '@cow/abis/types'
 import { BigNumber } from '@ethersproject/bignumber'
 import { ContractTransaction } from '@ethersproject/contracts'
-import { logSwapFlow, logSwapFlowError } from '@cow/modules/swap/services/swapFlow/logger'
+import { logSwapFlow, logSwapFlowError } from '@cow/modules/swap/services/utils/logger'
 
 // Use a 150K gas as a fallback if there's issue calculating the gas estimation (fixes some issues with some nodes failing to calculate gas costs for SC wallets)
 const PRESIGN_GAS_LIMIT_DEFAULT = BigNumber.from('150000')
